url_preview module : Fix random freeze while fetching URLs
* don't use requestfinished in puppeteer
inflight fails to reach 0 in some situation. I wasn't able to pinpoint the exact issue, but i'm assuming it's due to a race condition on the variable. This rework fix the issue by getting rid of inflight
Example of URL which sometimes fail in url_preview : https://stackoverflow.com/questions

* Extend time before screenshot, add some failsafe if the URL is unreachable or if folder don't exist